Add -i INTERVAL option to all checks
This fixes a UX error where await did not actually wait after failed
checks. The various *_check (mongdb_check etc) may specific a connection
timeout option. This defaults to 1 second. However, the connection
timeout does not ensure that checks themselves are run at that timeout.
Here's an example.

Use the mysql check against a working server but with incorrect
credentials. The supplied connection timeout would not matter because a
TCP connection establishes and the connection closes immediately due to
incorrect credentials. Subsequent checks execute immediately and fail
immediately.

This commit ensures that regardless of the underlying implementation in
the check functions there is a delay between each test.